LayoutTests/media/modern-media-controls/time-label/time-label-inherited-text-indent.html

<!DOCTYPE html>
<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<video src="../../content/test.mp4" style="width: 320px; height: 240px; text-indent: 100px;" controls autoplay></video>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing that the text-indent property set on the host does not affect time labels.");

const media = document.querySelector("video");
const shadowRoot = window.internals.shadowRoot(media);

shouldBecomeEqual("!!shadowRoot.querySelector('.time-label')", "true", () => {
    shouldBeEqualToString("getComputedStyle(shadowRoot.querySelector('.time-label')).textIndent", "0px");
    finishJSTest();
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
