Safari-609.1.11.1/LayoutTests/media/modern-media-controls/time-label/time-label-inherited-text-indent-expected.txt

Testing that the text-indent property set on the host does not affect time labels.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !!shadowRoot.querySelector('.time-label') became true
PASS getComputedStyle(shadowRoot.querySelector('.time-label')).textIndent is "0px"
PASS successfullyParsed is true

TEST COMPLETE
